Fix uint to be uint8 since otherwise it's undefined on a lot of platforms.

+++ BPTCEncoder/src/RGBAEndpoints.cpp
@@ -108,7 +108,7 @@ static inline void clamp(ty &x, const ty &min, const ty &max) {
 // absolute distance. It turns out the compiler does a much
 // better job of optimizing this than we can, since we can't 
 // translate the values to/from registers
-static uint sad(uint8 a, uint8 b) {
+static uint8 sad(uint8 a, uint8 b) {
 #if 0
 	__asm
 	{
